Update audio and video synchronization algorithm
Fix bugs

+++ VideoPlayer/src/player/videoplayer_p.h
@@ -33,7 +33,7 @@ public:
     bool isVideoInfoChanged = false;
 
     int position = 0;
-    int interval = 0;
+    qreal interval = 0;
 
     QQuickWindow* window() const { return q_ptr->window(); }
